openid-security: app work from docker-compose

+++ openid-security/idp/src/main/resources/idp_setup.sh
@@ -1,10 +1,9 @@
-# Use a prepackaged keycloak
-KEYCLOAK_IMAGE=quay.io/keycloak/keycloak:10.0.2
-
-#
 # Script to launch a keycloak IdP docker with a new realm
 # (admin UI: http://localhost:8080/auth/admin/)
 # See https://hub.docker.com/r/jboss/keycloak/
+
+# Can be used to preapre a docker-compose or launch a single IdP container
+# by adding argument -launch
 
 # settings for confbox ldap, access, tic
 # if ldap/web have real dns names you can use those instead of ip address
@@ -14,16 +13,24 @@ if [ -z "$WEB_HOST" -a -z "$WEB_URL" ]; then
 fi
 
 LDAP_HOST=${LDAP_HOST:-${WEB_HOST}}
-export WEB_URL=${WEB_URL:-http://${WEB_HOST}:8080/openid-security}
+export WEB_URL=${WEB_URL:-http://${WEB_HOST}:8080/openid-security-wildfly}
 export LDAP_URL=${LDAP_URL:-ldap://${LDAP_HOST}:389}
 
-echo "Patching realm file:"
+KEYCLOAK_REALM_FILE=/tmp/realm_data.json
+echo "Patching realm file: $KEYCLOAK_REALM_FILE"
 echo $LDAP_URL
 echo $WEB_URL
-KEYCLOAK_REALM_FILE=/tmp/realm_data.json
-cp idp_realm_auth_test.json ${KEYCLOAK_REALM_FILE}
-sed -i "s;http://WEB_HOST:8080/openid-security;${WEB_URL};g" ${KEYCLOAK_REALM_FILE}
-sed -i "s;ldap://LDAP_HOST:389;${LDAP_URL};g" ${KEYCLOAK_REALM_FILE}
+cp -f $(dirname $0)/idp_realm_auth_test.json ${KEYCLOAK_REALM_FILE}
+sed -i "s;http://.*/openid-security;${WEB_URL};g" ${KEYCLOAK_REALM_FILE}
+sed -i "s;ldap://.*:389;${LDAP_URL};g" ${KEYCLOAK_REALM_FILE}
+
+if [ "$1" != "-launch" ]; then
+    echo "Please run docker-compose up"
+    exit 0
+fi
+
+# Use a prepackaged keycloak
+KEYCLOAK_IMAGE=quay.io/keycloak/keycloak:10.0.2
 
 # Import predefined realm. Do not detach, log is useful for troubleshooting
 docker run --rm -p 8081:8080 --name keycloak_idp -e KEYCLOAK_USER=admin -e KEYCLOAK_PASSWORD=admin -e  KEYCLOAK_IMPORT=/tmp/realm_data.json -v ${KEYCLOAK_REALM_FILE}:/tmp/realm_data.json ${KEYCLOAK_IMAGE}
